Rework checkoutFrom handling and split project-clone into packages
Rework how project-clone handles checkoutFrom in a project to correctly
create a branch and split project-clone into separate files.
